JWT implemented

+++ backend/db/repository/blog.py
@@ -21,21 +21,26 @@ def list_blogs(db: Session):
     blogs=db.query(Blog).filter(Blog.is_active==True).all()
     return blogs
 
-def update_blog_by_id(id: int, blog:UpdateBlog, db:Session, author_id:int=1):
+def update_blog_by_id(id: int, blog:UpdateBlog, db:Session, author_id:int):
     blog_in_db=db.query(Blog).filter(Blog.id==id).first()
     if not blog_in_db:
-        return
+        return{"error":f"blog with {id} does not exists"}
+    if not blog_in_db.owner_id== author_id:
+        return {"error": f"Only the author can modify the blog"}
     blog_in_db.title=blog.title
     blog_in_db.content=blog.content
-    db.add(blog_in_db)
+    #db.add(blog_in_db)
     db.commit()
     db.refresh(blog_in_db)
     return blog_in_db
     
-def delete_by_id(id:int, db:Session,author_id:int=1):
+def delete_by_id(id:int, db:Session,author_id:int):
     blog_in_db=db.query(Blog).filter(Blog.id==id)
     if not blog_in_db.first():
         return {"error": f"Cound not find the blog with id {id}"}
+    if not blog_in_db.first().owner_id== author_id:
+        return {"error": "Only the author can delete the blog"}
+    
     blog_in_db.delete()
     db.commit()
     return {"msg": f"deleted the entry with id {id}"}
